+++ src/BaseClient.ts
@@ -7,11 +7,8 @@ import {CryptoGetAccountBalanceQuery} from "./generated/CryptoGetAccountBalance_
 import {QueryHeader} from "./generated/QueryHeader_pb";
 
 import {
-    accountIdToProto,
     handleQueryPrecheck,
-    normalizeAccountId,
     reqDefined,
-    tinybarRangeCheck
 } from "./util";
 import {ProtobufMessage} from "@improbable-eng/grpc-web/dist/typings/message";
 import {AccountCreateTransaction} from "./account/AccountCreateTransaction";
@@ -19,10 +16,10 @@ import {CryptoTransferTransaction} from "./account/CryptoTransferTransaction";
 import BigNumber from "bignumber.js";
 import {CryptoService} from "./generated/CryptoService_pb_service";
 
-import {Tinybar} from "./types/Tinybar";
+import {Tinybar, tinybarRangeCheck} from "./types/Tinybar";
 import {Hbar} from "./Hbar";
 import UnaryMethodDefinition = grpc.UnaryMethodDefinition;
-import {AccountId, AccountIdLike, accountIdToSdk} from "./types/AccountId";
+import {AccountId, AccountIdLike, accountIdToProto, accountIdToSdk, normalizeAccountId} from "./types/AccountId";
 import {TransactionId} from "./types/TransactionId";
 
 export type Signer = (msg: Uint8Array) => Uint8Array | Promise<Uint8Array>;
@@ -137,7 +134,7 @@ export abstract class BaseClient {
             .executeForReceipt()
             .then((receipt) => ({
                 account: accountIdToSdk(
-                    reqDefined(receipt.getAccountid(),
+                    reqDefined(accountIdToProto(receipt.accountId!),
                         'missing account ID from receipt: ' + receipt)),
             }));
     }
