+++ src/file/FileCreateTransaction.ts
@@ -8,7 +8,8 @@ import {FileService} from "../generated/FileService_pb_service";
 import {FileCreateTransactionBody} from "../generated/FileCreate_pb";
 import {Ed25519PublicKey} from "../Keys";
 import {KeyList} from "../generated/BasicTypes_pb";
-import {dateToTimestamp, timestampToProto} from "../util";
+import {timestampToProto} from "../util";
+import {dateToTimestamp} from "../types/Timestamp";
 
 export class FileCreateTransaction extends TransactionBuilder {
     private readonly body: FileCreateTransactionBody;
