+++ src/types/ContractFunctionResult.ts
@@ -6,7 +6,7 @@ export type ContractFunctionResult = {
     contractId: ContractId;
     contractCallResult: Uint8Array | string;
     errorMessage: string;
-    bloom: Uint8Array | string;
+    bloom: Uint8Array;
     gasUsed: number;
     logInfoList: ContractLogInfo[];
 }
@@ -16,7 +16,7 @@ export function contractFunctionResultToSdk(result: ProtoContractFunctionResult)
         contractId: contractIdToSdk(result.getContractid()!),
         contractCallResult: result.getContractcallresult(),
         errorMessage: result.getErrormessage(),
-        bloom: result.getBloom(),
+        bloom: result.getBloom_asU8(),
         gasUsed: result.getGasused(),
         logInfoList: contractLogInfoListToSdk(result.getLoginfoList())
     }
