+++ src/types/ExchangeRate.ts
@@ -1,15 +1,27 @@
-import {ExchangeRate as ProtoExchangeRate} from "../generated/ExchangeRate_pb";
+import {ExchangeRate as ProtoExchangeRate, ExchangeRateSet as ProtoExchangeRateSet} from "../generated/ExchangeRate_pb";
+
+export type ExchangeRateSet = {
+    currentRate: ExchangeRate;
+    nextRate: ExchangeRate;
+}
 
 export type ExchangeRate = {
     hbarEquiv: number;
     centEquiv: number;
-    expirationTime: number;
+    expirationTime: Date;
 }
 
 export function exchangeRateToSdk(exchangeRate: ProtoExchangeRate): ExchangeRate {
     return {
         hbarEquiv: exchangeRate.getHbarequiv(),
         centEquiv: exchangeRate.getCentequiv(),
-        expirationTime: exchangeRate.getExpirationtime()!.getSeconds()
+        expirationTime: new Date(exchangeRate.getExpirationtime()!.getSeconds())
+    }
+}
+
+export function exchangeRateSetToSdk(exchangeRateSet: ProtoExchangeRateSet): ExchangeRateSet {
+    return {
+        currentRate: exchangeRateToSdk(exchangeRateSet.getCurrentrate()!),
+        nextRate: exchangeRateToSdk(exchangeRateSet.getCurrentrate()!)
     }
 }
