+++ src/types/TransferList.ts
@@ -1,16 +1,8 @@
 import {AccountAmount, accountAmountToSdk} from "./AccountAmount";
 import {TransferList as ProtoTransferList} from "../generated/CryptoTransfer_pb";
 
-export type TransferList = {
-    accountAmountsList: AccountAmount[];
-}
-
-export function transferListToSdk(transferList: ProtoTransferList): TransferList {
-    const accountAmountsList = transferList.getAccountamountsList().map((accountAmount) => {
+export function transferListToSdk(transferList: ProtoTransferList): AccountAmount[] {
+    return transferList.getAccountamountsList().map((accountAmount) => {
         return accountAmountToSdk(accountAmount);
     });
-
-    return {
-        accountAmountsList
-    }
 }
